Improve ranking configurability and performance

+++ code/functions/language-processing/java/nu/marginalia/keyword/model/DocumentKeywords.java
@@ -5,7 +5,7 @@ import nu.marginalia.sequence.VarintCodedSequence;
 import java.util.List;
 
 public record DocumentKeywords(List<String> keywords,
-                               byte[] metadata,
+                               long[] metadata,
                                List<VarintCodedSequence> positions,
                                byte[] spanCodes,
                                List<VarintCodedSequence> spanSequences) {
